add fee column and refactor ui
