+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["crates/tel-core", "crates/tel-api", "crates/tel-ui", "crates/tel-indexer"]
+members = ["crates/tel-core", "crates/tel-api", "crates/tel-ui", "crates/tel-indexer", "bin/tel-ui"]
 resolver = "2"
