+++ bin/tel-ui/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tel-ui-bin"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tel-ui = { path = "../../crates/tel-ui" }
+eframe = "0.27.2"
+egui = "0.27.2"
+tracing-subscriber = "0.3.18"
+
+[[bin]]
+name = "tel-ui"
+path = "main.rs"
