+++ crates/tel-indexer/Cargo.toml
@@ -12,7 +12,7 @@ tel-core = { path = "../tel-core" }
 tokio = { version = "1.31.0", features = ["full"] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-alloy-primitives = "1.1.0"
+alloy-primitives = "1.0.1"
 anyhow = "1.0.75"
 clap = { version = "4.4.6", features = ["derive"] }
