+++ crates/tel-indexer/src/lib.rs
@@ -1,4 +1,3 @@
-use crate::storage::Storage;
 use alloy_primitives::Address;
 use std::collections::HashMap;
 use std::str::FromStr;
@@ -11,6 +10,7 @@ use tel_core::models::{LiquidityDistribution, Pool, Token};
 use tel_core::providers::ProviderManager;
 use tel_core::storage;
 use tel_core::storage::SqliteStorage;
+use tel_core::storage::Storage;
 use tokio::time;
 use tracing::{debug, error, info, warn};
