+++ crates/tel-ui/Cargo.toml
@@ -10,9 +10,9 @@ license = "MIT"
 tel-core = { path = "../tel-core" }
 
 # UI dependencies
-egui = "0.26.2"
-eframe = { version = "0.26.2", features = ["default_fonts", "persistence"] }
-egui_plot = "0.26.2"
+egui = "0.27.2"
+eframe = { version = "0.27.2", features = ["default_fonts", "persistence"] }
+egui_plot = "0.27.2"
 poll-promise = "0.3.0"
 tokio = { version = "1.31.0", features = ["full"] }
 serde = { version = "1.0.188", features = ["derive"] }
@@ -22,8 +22,8 @@ rusqlite = { version = "0.28.0", features = ["bundled"] }
 chrono = { version = "0.4.31", features = ["serde"] }
 tracing = "0.1.37"
 tracing-subscriber = "0.3.17"
-alloy-primitives = "1.1.0"  # Same version as tel-core
+alloy-primitives = "1.0.1"  # Same version as tel-core
 
-[[bin]]
-name = "tel-ui"
-path = "../../bin/tel-ui/main.rs" 
+[lib]
+name = "tel_ui"
+path = "src/lib.rs"
